% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcub.R
\name{bcub}
\alias{bcub}
\title{bcub}
\usage{
bcub(pi.fo, xi.fo, m, shift = 1, data = NULL, pi.link = "probit",
  xi.link = "probit", N = 5000)
}
\arguments{
\item{pi.fo}{an object of class \code{\link[stats]{formula}}, a symbolic description of the model to fit pi parameter.}

\item{xi.fo}{an object of class \code{\link[stats]{formula}}, a symbolic description of the model to fit xi parameter without left side.}

\item{m}{the maximum value of the response variable.}

\item{shift}{the minimum value of the response variable.}

\item{data}{an optional data frame.}

\item{pi.link}{link function to use for model pi parameter, by default is probit.}

\item{xi.link}{link function to use for model xi parameter, by default is probit.}

\item{N}{This required argument accepts integers larger than 10, and determines the number of iterations that Laplace's Demon will update the parameters while searching for target distributions.}
}
\description{
Function to fit a cub model under a bayesian approach. The structure can have covariates or not to model \eqn{\pi} and \eqn{\xi} parameters.
}
\examples{

1 + 3

}
