% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cub.R
\name{boot.cub}
\alias{boot.cub}
\title{Bootstrap
This function is used to obtain standard error for betas
by bootstrap method.}
\usage{
boot.cub(object)
}
\description{
Bootstrap
This function is used to obtain standard error for betas
by bootstrap method.
}
