% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condom.R
\docType{data}
\name{condom}
\alias{condom}
\title{Frequency of condom use in 2013}
\format{A data frame with 153 rows and 4 variables:
\describe{
  \item{age}{age, in years}
  \item{gender}{A factor with levels: 0 = woman, 1 = man}
  \item{stratum}{socioeconomic level}
  \item{usecondom}{A factor with levels: 1 = always, 2 = frequently, 3 = occasionally and 4 = never}
}}
\usage{
data(condom)
}
\description{
A dataset containing the frequency of condom use in Medellin in 2013.
}
\keyword{datasets}
