% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cub.R
\name{cub}
\alias{cub}
\alias{summary.cub}
\title{cub}
\usage{
cub(
  pi.fo,
  xi.fo,
  m,
  data = NULL,
  subset = NULL,
  optimizer = "nlminb",
  pi.link = "probit",
  xi.link = "probit",
  initial.values = NULL,
  ...
)
}
\arguments{
\item{pi.fo}{an object of class \code{\link[stats]{formula}}, a symbolic description of the model to fit pi parameter.}

\item{xi.fo}{an object of class \code{\link[stats]{formula}}, a symbolic description of the model to fit xi parameter without left side.}

\item{m}{the maximum value of the response variable.}

\item{data}{an optional data frame.}

\item{subset}{an optional vector specifying a subset of observations to be used. For example, \code{subset = "sex =='male'"} means that we want to use only the male data set.}

\item{optimizer}{two options are available: \code{\link[stats]{nlminb}} (default), \code{\link[stats]{optim}} or \code{\link[DEoptim]{DEoptim}}.}

\item{pi.link}{link function to use for model pi parameter, two options are available, logit or probit, by default is probit.}

\item{xi.link}{link function to use for model xi parameter, two options are available, logit or probit, by default is probit.}

\item{initial.values}{vector with the initial values to the searching procedure to nlminb and optim optimizers.}

\item{...}{Further arguments to be passed to \code{\link[DEoptim]{DEoptim.control}}.}
}
\description{
function to fit a cub model. The user could use covariates or not to model the parameters pi and xi.
}
\examples{
# Example 1 ---------------------------------------------------------------
# Generating a random sample given the values of pi and xi
set.seed(2019)
y <- rcub(n=100, pi=0.15, xi=0.60, m=5)
mod1 <- cub(pi.fo=y ~ 1, xi.fo=~ 1, m=5,
            pi.link='probit', xi.link='probit', 
            optimizer='nlminb')
summary(mod1)

# estimations for pi and xi
pnorm(mod1$par[1])
pnorm(mod1$par[2])

# Example 2 ---------------------------------------------------------------
# Generating a random sample given the values of pi and xi 
# estimation using logit link function
set.seed(2019)
y <- rcub(n=100, pi=0.15, xi=0.60, m=5)
mod2 <- cub(pi.fo=y ~ 1, xi.fo=~ 1, m=5, 
            pi.link='logit', xi.link='logit', 
            optimizer='nlminb')
summary(mod2)

# estimations for pi and xi
boot::inv.logit(mod2$par[1])
boot::inv.logit(mod2$par[2])

# Example 3 ---------------------------------------------------------------
# Generating a random sample given the values of pi and xi 
# estimation using logit and probit as link function
set.seed(2019)
y <- rcub(n=100, pi=0.15, xi=0.60, m=5)
mod3 <- cub(pi.fo=y ~ 1, xi.fo=~ 1, m=5, 
            pi.link='logit', xi.link='probit',
            optimizer='nlminb')
summary(mod3)
# estimations for pi and xi
boot::inv.logit(mod3$par[1])
pnorm(mod3$par[2])

}
