% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F2.R
\name{F2.cub}
\alias{F2.cub}
\title{F2 coefficient for cub model}
\usage{
F2.cub(mod)
}
\arguments{
\item{mod}{object of cub class}
}
\description{
This function generates the F2 coefficient, difference 1 and Diss.
}
\examples{

# Test 1 

# Generating a random sample given the values of pi and xi

y <- rcub(n = 1000, pi = 0.15, xi = 0.60, m = 8)
mod <- cub(pi.fo = y ~ 1, xi.fo = ~ 1, m = 8, shift = 1)
F2.cub(mod)

# Test 2 

fit <- cub(pi.fo = usecondom ~ gender, xi.fo = ~ gender,  m = 4, data = data)
F2.cub(fit)

# Test 3 

# Simulating a dataset with qualitative variables
n <- 10000
# Betas for pi
b0 <- -1
b1 <- 1
# Betas for xi
g0 <- -2
g1 <- 1.5

beta_japan <- 1 # Coefficient for japan
beta_usa <- -0.7 # Coefficient for usa, france is reference level

# m and shift value
m <- 5
shift <- 1

# Simulating the dataset
x1 <- runif(n)
x2 <- runif(n)
country <- sample(c('france', 'japan', 'usa'), size = n, replace = T)
aux <- model.matrix(~country)[, -1] \%*\% matrix(c(beta_japan,  beta_usa))
pi <- pnorm(b0 + b1 * x1 + aux)
xi <- pnorm(g0 + g1 * x2 + aux)
y <- rcub(n = n, pi = pi, xi = xi, m = m, shift = shift)
# Dataset
dataset <- data.frame(y, x1, x2, country)

# Fitting the model
mod <- cub(pi.fo = y ~ x1 + country,
           xi.fo = ~ x2 + country, m = 5, data = dataset,
           optimizer = 'optim')
F2.cub(mod)

}
