% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icub.R
\name{icub}
\alias{icub}
\title{I coefficient for cub model without covariables}
\usage{
icub(mod)
}
\arguments{
\item{mod}{object of cub class.}
}
\description{
This function generates I coefficient.
}
\examples{

# Test 1 

# Generating a random sample given the values of pi and xi

y <- rcub(n=1000, pi=0.2, xi=0.70, m=5)
mod <- cub(pi.fo=y ~ 1, xi.fo=~ 1, m=5)
icub(mod)

y <- rcub(n=100, pi=0.9, xi=0.5, m=10)
mod1 <- cub(pi.fo=y ~ 1, xi.fo=~ 1, m=10)
icub(mod1)

}
