% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcub.R
\name{pcub}
\alias{pcub}
\title{Distribution function for cub model}
\usage{
pcub(q, pi, xi, m, shift = 1, lower.tail = TRUE, log = FALSE)
}
\arguments{
\item{q}{vector of quantiles.}

\item{pi}{uncertainty parameter belongs to \code{(0, 1]}}

\item{xi}{feeling parameter belongs to \code{[0, 1]}}

\item{m}{the maximum value for random values}

\item{shift}{the minimum value for random values}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \leq x]} otherwise, \eqn{P[X > x]}.}

\item{log}{logical; if TRUE, densities are given as log.}
}
\description{
This function gives the probability distribution of a cub model without covariates given the parameters \eqn{\pi} and \eqn{\xi}.
}
\examples{
# low xi is associated with high ratings
pcub(q=5, pi=0.5, xi=0.2, m=10, lower.tail=FALSE)

# high pi is associated with indecision in choosing
pcub(q=3, pi=0.9, xi=0.5, m=4) 

# probability for several quantiles
pcub(q=c(1,3,5), pi=0.3, xi=0.6, m=5)

}
