% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cub.R
\name{print.cub}
\alias{print.cub}
\title{Print cub class
This function is used to print an object of class cub.}
\usage{
\method{print}{cub}(mod, ...)
}
\description{
Print cub class
This function is used to print an object of class cub.
}
