% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcub.R
\name{qcub}
\alias{qcub}
\title{Quantile function for cub model}
\usage{
qcub(p, pi, xi, m, shift = 1, lower.tail = TRUE, log = FALSE)
}
\arguments{
\item{p}{vector of probabilities.}

\item{pi}{uncertainty parameter belongs to \code{(0, 1]}}

\item{xi}{feeling parameter belongs to \code{[0, 1]}}

\item{m}{the maximum value for random values}

\item{shift}{the minimum value for random values}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \leq x]} otherwise, \eqn{P[X > x]}.}

\item{log}{logical; if TRUE, densities are given as log.}
}
\description{
This function gives the quantile distribution of a cub model without covariates given the parameters \eqn{\pi} and \eqn{\xi}.
}
\examples{

# low xi is associated with high ratings
qcub(p=0.1, pi=0.5, xi=0.1, m=7, lower.tail=T)

# high pi is associated with indecision in choosing
qcub(p=0.86, pi=0.9, xi=0.5, m=4)

#quantiles for several probabilities 
qcub(p=c(1,0.5,0.9), pi=0.3, xi=0.6, m=5)

}
