% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcub.R
\name{rcub}
\alias{rcub}
\title{Random samples from a cub model}
\usage{
rcub(n, pi, xi, m = 5, shift = 1)
}
\arguments{
\item{n}{number of observations}

\item{pi}{uncertainty parameter belongs to \code{(0, 1]}}

\item{xi}{feeling parameter belongs to \code{[0, 1]}}

\item{m}{the maximum value for random values}

\item{shift}{the minimum value for random values}
}
\description{
This function generates random samples from a cub model given the parameters \eqn{\pi} and \eqn{\xi}.
}
\examples{
# Random sample, low xi is associated with high ratings
x <- rcub(n=1000, pi=0.9, xi=0.1, m=5)
barplot(prop.table(table(x)))

# Random sample, low pi is associated with random choices
y <- rcub(n=1000, pi=0.1, xi=0.1, m=5)
barplot(prop.table(table(y)))

}
