% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{rq.res}
\alias{rq.res}
\title{Randomized Quantile Residuals for cub models.}
\usage{
rq.res(mod)
}
\arguments{
\item{mod}{object of cub class.}
}
\description{
This function generates Randomized Quantile Residuals.
}
\examples{

# Test 1

y <- rcub(n=1000, pi=0.1, xi=0.70, m=8)
mod0 <- cub(pi.fo=y ~ 1, xi.fo=~1, m=8)
residuos <- rq.res(mod0)
qqnorm(residuos, ylab='Cuantiles muestrales',
       xlab='Cuantiles teoricos',las=1)
qqline(residuos, col='red')

# Test 2

rcub.covariates <- function(n, b0, b1, g0, g1, m = 5, shift = 1) {
x1 <- runif(n)
x2 <- runif(n)
pi <- pnorm(b0 + b1 * x1)
xi <- pnorm(g0 + g1 * x2)
y <- rcub(n=n, pi=pi, xi=xi, m=m)
data.frame(y, x1, x2, xi, pi)}

# Generating the data
dataset <- rcub.covariates(n=100, b0=-1, b1=1, g0=-2, g1=1.5)

# Fitting the model with optim optimizer
fit1 <- cub(pi.fo=y ~ x1, xi.fo=~ x2, m=5, data=dataset,
            optimizer='optim')
rq.res(fit1)

}
