% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update1.cub.R
\name{update1}
\alias{update1}
\title{Update and Re-fit a Model Call}
\usage{
update1(object, ..., evaluate = TRUE)
}
\arguments{
\item{object}{an existing fit obtained from cub function.}

\item{...}{additionally arguments to the call or arguments with changed values.}

\item{evaluate}{if true evaluate the new call else return the call.}
}
\description{
This function can be used to fit a new cub model in which has changed some arguments.
}
\examples{
# The first model
mod1 <- cub(pi.fo = global ~ 1, xi.fo = ~ 1, m=7, data=univer)
mod1
# Modifying right sides of pi and xi formulas
update1(mod1, pi.fo = global ~ gender, xi.fo = ~ lage + gender)

# Modifying the dataset
update1(mod1, xi.fo = ~ 1, m=7, data=univer[1:50, ])

}
