% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update2.cub.R
\name{update2.cub}
\alias{update2.cub}
\title{update2.cub}
\usage{
update2.cub(object, ..., evaluate = TRUE)
}
\arguments{
\item{n}{number of observations}

\item{pi}{uncertainty parameter belongs to \code{(0, 1]}}

\item{xi}{feeling parameter belongs to \code{[0, 1]}}

\item{m}{the maximum value for random values}

\item{shift}{the minimum value for random values}
}
\description{
Update and Re-fit a Model Call
}
\examples{
# The first model
mod1 <- cub(pi.fo = global ~ 1, xi.fo = ~ 1, m=7, data=univer)
# Modifying right sides of pi and xi formulas
mod2 <- update2(mod1, pi.fo = global ~ gender, xi.fo = ~ lage + gender)
# Modifying the dataset
mod3 <- update2(xi.fo = ~ 1, m=7, data=univer[1:50, ])

}
