% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cub.R
\name{cub}
\alias{cub}
\alias{summary.cub}
\title{cub}
\usage{
cub(pi.fo, xi.fo, m, data = NULL, subset = NULL,
  optimizer = "nlminb", pi.link = "probit", xi.link = "probit",
  initial.values = NULL, ...)
}
\arguments{
\item{pi.fo}{an object of class \code{\link[stats]{formula}}, a symbolic description of the model to fit pi parameter.}

\item{xi.fo}{an object of class \code{\link[stats]{formula}}, a symbolic description of the model to fit xi parameter without left side.}

\item{m}{the maximum value of the response variable.}

\item{data}{an optional data frame.}

\item{subset}{an optional vector specifying a subset of observations to be used. For example, \code{subset = "sex =='male'"} means that we want to use only the male data set.}

\item{optimizer}{two options are available: \code{\link[stats]{nlminb}} (default), \code{\link[stats]{optim}} or \code{\link[DEoptim]{DEoptim}}.}

\item{pi.link}{link function to use for model pi parameter, two options are available, logit or probit, by default is probit.}

\item{xi.link}{link function to use for model xi parameter, two options are available, logit or probit, by default is probit.}

\item{initial.values}{vector with the initial values to the searching procedure to nlminb and optim optimizers.}

\item{...}{Further arguments to be passed to \code{\link[DEoptim]{DEoptim.control}}.}
}
\description{
function to fit a cub model. The user could use covariates or not to model the parameters \eqn{\pi} and \eqn{\xi}.
}
\examples{
x <- 1:5
y <- rnorm(5)
mod <- lm(y ~ x)

}
