% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cub.R
\name{cub}
\alias{cub}
\title{cub}
\usage{
cub(pi.fo, xi.fo, m, shift = 1, data = NULL, optimizer = "nlminb")
}
\arguments{
\item{pi.fo}{an object of class \code{\link[stats]{formula}}, a symbolic description of the model to fit pi parameter.}

\item{xi.fo}{an object of class \code{\link[stats]{formula}}, a symbolic description of the model to fit xi parameter without left side.}

\item{m}{the maximum value of the response variable.}

\item{shift}{the minimum value of the response variable.}

\item{data}{an optional data frame.}

\item{optimizer}{two options are available: \code{\link[stats]{nlminb}} (default) or \code{\link[stats]{optim}}.}
}
\description{
Function to fit a cub model. The structure can have covariates or not to model \eqn{\pi} and \eqn{\xi} parameters.
}
\examples{

# Test 1 ------------------------------------------------------------------
# Generating a random sample given the values of pi and xi
y <- rcub(n=10000, pi=0.15, xi=0.60, m=5)
mod <- cub(pi.fo = y ~ 1, xi.fo = ~ 1, m=5, shift=1,
           optimizer='nlminb')
summary(mod)
# To obtain the fitted parameters in the appropiate scale
# we need to apply the inverse link function to the fitted parameters
pnorm(mod$par)


# Test 2 ------------------------------------------------------------------
# rcub.covariates is a function to generate a random sample from a cub model 
# using covariates to model pi and xi and link function qnorm
rcub.covariates <- function(n, b0, b1, g0, g1, m = 5, shift = 1) {
  x1 <- runif(n)
  x2 <- runif(n)
  pi <- pnorm(b0 + b1 * x1)
  xi <- pnorm(g0 + g1 * x2)
  y <- rcub(n = n, pi = pi, xi = xi, m = m, shift = shift)
  data.frame(y, x1, x2)
}

# Generating the data
dataset <- rcub.covariates(n=1000, b0=-1, b1=1, g0=-2, g1=1.5)
# Fitting the model
mod <- cub(pi.fo = y ~ x1, xi.fo = ~ x2, m=5, data=dataset,
           optimizer='optim')
summary(mod)

# Test 3 ------------------------------------------------------------------
# Simulating a dataset with qualitative variables
n <- 1000
# Betas for pi
b0 <- -1
b1 <- 1
# Betas for xi
g0 <- -2
g1 <- 1.5

beta_japan <- 1 # Coefficient for japan
beta_usa <- 1.5 # Coefficient for usa, france is reference level

# m and shift value
m <- 5
shift <- 1

# Simulating the dataset
x1 <- runif(n)
x2 <- runif(n)
country <- sample(c('france', 'japan', 'usa'), size=n, replace=T)
aux <- model.matrix(~country)[, -1] \%*\% matrix(c(beta_japan,  beta_usa))
pi <- pnorm(b0 + b1 * x1 + aux)
xi <- pnorm(g0 + g1 * x2 + aux)
y <- rcub(n = n, pi = pi, xi = xi, m = m, shift = shift)
# Dataset
dataset <- data.frame(y, x1, x2, country)

# Fitting the model
mod <- cub(pi.fo = y ~ x1 + country,
           xi.fo = ~ x2 + country, m=5, data=dataset,
           optimizer='optim')
summary(mod)

}

