% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcub.R
\name{Cub}
\alias{Cub}
\alias{dcub}
\alias{pcub}
\alias{qcub}
\alias{rcub}
\title{cub distribution}
\usage{
dcub(x, pi, xi, m, log = FALSE, ...)

pcub(q)

qcub(p)

rcub(n)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{pi}{uncertainty parameter belongs to \code{(0, 1]}.}

\item{xi}{feeling parameter belongs to \code{[0, 1]}.}

\item{m}{the maximum value.}

\item{log}{logical; if TRUE, densities are given as log.}

\item{p}{vector of probabilities.}

\item{n}{number of observations}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X ≤ x]} otherwise, \code{P[X > x]}.}
}
\description{
Density, distribution function, quantile function and random generation for the cub distribution given parameters \eqn{\pi} and \eqn{\xi}.
}
\examples{
dcub(x=4, pi=0.3, xi=0.7, m=5)
dcub(x=1, pi=0.5, xi=0.4, m=8)
dcub(x=c(4, 1), pi=c(0.3, 0.5), xi=c(0.7, 0.4), m=c(5, 8))
}
