% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diss.R
\name{Diss}
\alias{Diss}
\title{The dissimilarity index Diss for the cub model}
\usage{
Diss(mod)
}
\arguments{
\item{mod}{object of class \code{"cub"}.}
}
\description{
This function generates the Diss coefficient, the difference between observed relative frequencies and estimated (theoretical) probabilities.
}
\examples{
# Test 1 
y <- rcub(n=800, pi=0.30, xi=0.80, m=6)
fit1 <- cub(pi.fo=y ~ 1, xi.fo= ~ 1, m=6)
Diss(fit1)

# Test 2
fit2 <- cub(pi.fo=global ~ gender + lage,
            xi.fo= ~ residenc + willingn ,
            m=7, data=univer)
Diss(fit2)

}
