% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqr.R
\name{rqr}
\alias{rqr}
\title{Randomized Quantile Residuals for cub models.}
\usage{
rqr(y, pi, xi, m)
}
\arguments{
\item{y}{vector with the response variable.}

\item{pi}{value or vector with the values for pi parameter.}

\item{xi}{value or vector with the values for pi parameter.}

\item{m}{the maximum value.}
}
\description{
This function generates Randomized Quantile Residuals.
}
\examples{
y <- rcub(n=1000, pi=0.15, xi=0.60, m=8)
# To fit the model
mod1 <- cub(pi.fo=y~1, xi.fo=~1, m=8)
# To obtain the RQR
r1 <- rqr(y, pi=mod1$fitted.pi, xi=mod1$fitted.xi, m=8)
# To check the normality
qqplot(y=r1, x=qnorm(ppoints(1000), mean=0, sd=1), las=1,
       ylab='RQR', xlab='Theoretical Quantiles N(0, 1)')
abline(a=0, b=1, col='red')
# Using a test to check if rqr follow a N(0, 1)
ks.test(r1, "pnorm", mean=0, sd=1)

}
